共有URLのメタデータ
